+++ src/org/digitalcampus/oppia/task/DownloadCourseTask.java
@@ -90,8 +90,8 @@ public class DownloadCourseTask extends AsyncTask<Payload, DownloadProgress, Pay
 							ctx.getString(R.string.prefServerTimeoutResponse))));
             
 			
-			int fileLength = c.getContentLength();
-            int availableStorage = FileUtils.getAvailableStorageSize(ctx);
+			long fileLength = c.getContentLength();
+            long availableStorage = FileUtils.getAvailableStorageSize(ctx);
 
             if (fileLength >= availableStorage){
                 payload.setResult(false);
@@ -113,7 +113,7 @@ public class DownloadCourseTask extends AsyncTask<Payload, DownloadProgress, Pay
                 int progress = 0;
                 while ((len1 = in.read(buffer)) > 0) {
                     total += len1;
-                    progress = (int)(total*100)/fileLength;
+                    progress = (int)((total*100)/fileLength);
                     if(progress > 0){
                         dp.setProgress(progress);
                         publishProgress(dp);
